+++ superwall/src/main/java/com/superwall/sdk/paywall/vc/web_view/templating/TemplateLogic.kt
@@ -26,7 +26,7 @@ object TemplateLogic {
         )
 
         val variablesTemplate = factory.makeJsonVariables(
-            productVariables = paywall.productVariables,
+            products = paywall.productVariables,
             computedPropertyRequests = paywall.computedPropertyRequests,
             event = event
         )